MB-36249: Don't floor() write amplification stats
Fix calculation of ep_io_flusher_write_amplification &
ep_io_total_write_amplification to not crop the fractional part - they
should be floating-point values.

Reviewed-on: http://review.couchbase.org/117512
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/ep_engine.cc
@@ -2991,7 +2991,7 @@ ENGINE_ERROR_CODE EventuallyPersistentEngine::doEngineStats(
                 kvBucket->getKVStoreStat(writeBytesStat,
                                          bytesWritten,
                                          KVBucketIface::KVSOption::RW)) {
-                writeAmp = bytesWritten / docBytes;
+                writeAmp = double(bytesWritten) / docBytes;
             }
             add_casted_stat(writeAmpStat, writeAmp, add_stat, cookie);
         };
